requirements.txt
click==5.1
colorama==0.3.3
dj-database-url==0.4.1
Django==1.8
django-appconf==1.0.1
django-bootstrap-form==3.2
django-braces==1.8.1
django-classy-tags==0.7.0
django-cors-headers==1.1.0
django-grappelli==2.7.3
django-jsonfield==0.9.15
django-leaflet==0.18.0
django-model-utils==2.4
django-rest-auth==0.7.0
django-webpack-loader==0.2.4
djangorestframework==3.3.2
easypost==3.1.0
ecdsa==0.13
Fabric==1.10.2
googlemaps==2.4.2
gunicorn==19.4.5
httpie==0.9.3
jsonfield==1.0.3
lxml==3.6.0
metron==1.3.7
mysql==0.0.1
MySQL-python==1.2.5
paramiko==1.16.0
psycopg2==2.6.1
Pygments==2.1
pytz==2015.7
refreshbooks==2.0
requests==2.7.0
six==1.10.0
stripe==1.30.0
twitter==1.17.1
whitenoise==3.0
